py/requirements.txt
fastapi
uvicorn
python-socketio[asgi]
transformers